+++ app/src/main/java/woowacourse/paint/customview/Stroke.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.customview
 
 import android.graphics.Paint
 import android.graphics.Path
