+++ app/src/main/java/woowacourse/paint/main/ColorClickListener.kt
@@ -1,4 +1,6 @@
-package woowacourse.paint
+package woowacourse.paint.main
+
+import woowacourse.paint.main.model.BrushColorBox
 
 interface ColorClickListener {
     fun onColorClick(clickedBrushColorBox: BrushColorBox)
